requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAi Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic 

# Google Gemini (PaLM) Integration
langchain-google-genai
google.generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Enviroment Variable Management
python-dotenv

# Machine learning Utilites
numpy 
scikit-learn 

# for Data validation
pydantic
pydantic[email]

grandalf